Move util methods into rpc prov
